requirements_test.txt
pytest-asyncio==0.20.2
pytest-homeassistant-custom-component==0.12.48
